+++ CesiumMetadata/include/CesiumMetadata/MetadataConversions.h
