+++ CesiumMetadata/include/CesiumMetadata/PropertyArrayView.h
