+++ CesiumMetadata/include/CesiumMetadata/PropertyAttributeView.h
