+++ CesiumMetadata/include/CesiumMetadata/PropertyTableView.h
