+++ CesiumMetadata/include/CesiumMetadata/PropertyTexturePropertyView.h
