+++ CesiumMetadata/include/CesiumMetadata/PropertyTextureView.h
