+++ CesiumMetadata/include/CesiumMetadata/PropertyTransformations.h
