+++ CesiumMetadata/include/CesiumMetadata/PropertyTypeTraits.h
