+++ CesiumMetadata/include/CesiumMetadata/PropertyView.h
