+++ CesiumMetadata/test/TestMetadataConversions.cpp
