+++ CesiumMetadata/test/TestPropertyTexturePropertyView.cpp
