+++ CesiumMetadata/test/TestPropertyTextureView.cpp
