+++ CesiumMetadata/test/TestPropertyType.cpp
