+++ CesiumMetadata/test/TestPropertyTypeTraits.cpp
