+++ CesiumMetadata/test/makeEnumValue.cpp
