+++ CesiumMetadata/test/makeEnumValue.h
